feat(basic-display-api): add support for reading current user media
